cleanup connection pool return
This change ensures that connections that fail to return to the pool are
cleanly closed and exception raised are not returned to the caller.

For rabbit, we also try to reconnection in case of connection failure,
before dropping the connection.

Closes-bug: #1433458

Change-Id: Ic714db7b8be9df8b6935a903732c60aaea0bc404
(cherry picked from commit 0dff20b8b979d902d855aa1c0ae0f4b9398a3116)

Conflicts:
	oslo_messaging/tests/drivers/test_impl_rabbit.py