Don't reply when we known that client is gone
In case of a broker restart/failover a reply queue can be unreachable for short period the IncomingMessage.send_reply will block for 60 seconds in this case or until rabbit recovers.
But in case of the reply queue is unreachable because the rpc client is really gone, we can have a ton of reply to send waiting 60 seconds.
This leads to a starvation of connection of the pool The rpc server take to much time to send reply, other rpc client will raise TimeoutError because their don't receive their replies in time.
This changes introduces an object cache that stores already known gone client to not wait 60 seconds and hold a connection of the pool Keeping 200 last gone rpc client for 1 minute is enough and doesn't hold to much memory.
This also don't raise anymore a frightening exception when we can't send reply to the rpc client. But just logging a info about missing exchange and a warning about unsend reply.

Related upstream commit: 286659a38be5db399d0b9f807fac7b980d6c0b7e
modified the log message format in direct_send function

Signed-off-by: blkart <blkart.org@gmail.com>