Don't hold the connection when reply fail
This change moves the reply retry code to upper layer to be able to release the connection while we wait between two retries.

In the worse scenario, a client waits for more than 30 replies and died/restart, the server tries to send this 30 replies to this th
lient and can wait too 60s per replies. During this replies for other clients are just stuck.

This change fixes that.

Related-bug: #1477914
Closes-bug: #1521958

Related upstream commit: daddb82788918296f8b34d6cdeb40d01620fb183

Signed-off-by: blkart <blkart.org@gmail.com>